feat: fix erros of square int overflow and block capacity set before filling a new column

+++ source/libs/function/test/udf2.c
@@ -44,7 +44,7 @@ int32_t udf2(SUdfDataBlock* block, SUdfInterBuf *interBuf, SUdfInterBuf *newInte
         case TSDB_DATA_TYPE_INT: {
           char* cell = udfColDataGetData(col, j);
           int32_t num = *(int32_t*)cell;
-          sumSquares += num * num;
+          sumSquares += (double)num * num;
           break;
         }
         case TSDB_DATA_TYPE_DOUBLE: {
